refactor: 레이아웃 회원가입 버튼 추가
