+++ src/components/layout/navbar/Navbar.jsx
@@ -1,5 +1,5 @@
 import React, {useContext} from 'react';
-import { AuthContext } from '../../contexts/AuthProvider';
+import { AuthContext } from '../../../contexts/AuthProvider';
 import { useNavigate, Link } from 'react-router-dom';
 import './Navbar.css';
 
@@ -17,7 +17,7 @@ function Navbar() {
     }
 
     const signUpClick = () => {
-        navigate('/singup');
+        navigate('/signup');
     }
 
     return (
@@ -33,8 +33,10 @@ function Navbar() {
                         <button onClick={logoutClick} className="nav-button">로그아웃</button>
                     </>
                 ) : (
-                    <button onClick={loginClick} className={"nav-button"}>로그인</button>
-                    // <button onClick={signUpClick} className={"nav-button"}>회원가입</button>
+                    <>
+                        <button onClick={loginClick} className={"nav-button"}>로그인</button>
+                        <button onClick={signUpClick} className={"nav-button"}>회원가입</button>
+                    </>
                 )}
             </div>
         </header>
